requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
APScheduler==3.10.4
scikit-learn==1.5.0
Pillow==10.3.0
numpy==1.26.4
